+++ src/tdastro/consts.py
@@ -1,0 +1,3 @@
+from astropy import units as u
+
+PARSEC_TO_CM = (1 * u.pc).to_value(u.cm)
