+++ src/tdastro/sources/periodic_variable_star.py
@@ -1,12 +1,10 @@
 from abc import ABC, abstractmethod
 
-import astropy.units as u
 import numpy as np
-from astropy.modeling.physical_models import BlackBody
 
+from tdastro.astro_utils.black_body import black_body_luminosity_density
+from tdastro.consts import PARSEC_TO_CM
 from tdastro.sources.periodic_source import PeriodicSource
-
-PARSEC_TO_CM = (1 * u.pc).to_value(u.cm)
 
 
 class PeriodicVariableStar(PeriodicSource, ABC):
@@ -216,28 +214,3 @@ class EclipsingBinaryStar(PeriodicVariableStar):
         area = np.where(d <= minimum_radius, np.pi * minimum_radius**2, area)
 
         return area
-
-
-# We should move this function to a more appropriate location
-def black_body_luminosity_density(temperature, radius, wavelengths):
-    """Calculate the black-body luminosity density for a star.
-
-    Parameters
-    ----------
-    temperature : `float`
-        The effective temperature of the star, in kelvins.
-    radius : `float`
-        The radius of the star, in solar radii.
-    wavelengths : `numpy.ndarray`
-        A length N array of wavelengths.
-
-    Returns
-    -------
-    luminosity_density : `numpy.ndarray`
-        A length N array of luminosity density values.
-    """
-    black_body = BlackBody(temperature)
-    intensity_per_freq = black_body(wavelengths * u.cm).to_cgs().value
-    surface_flux = intensity_per_freq * np.pi
-    surface_area = 4.0 * np.pi * radius**2
-    return surface_flux * surface_area
